requirements.txt
python-telegram-bot>=20.0
openai>=0.27.0
python-dotenv>=1.0.0
redis>=4.5.0
pytest>=7.0.0